# backend/OPENAPI_SPEC.yaml
openapi: 3.0.3
info:
  title: Directory Wizard API
  description: |
    Multi-tenant Directory Creation Wizard API for creating and managing directory websites
    through a guided interface with integrated external service provisioning.
  version: 1.0.0
  contact:
    name: Directory Wizard Support
    email: support@directorywizard.com
    url: https://docs.directorywizard.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.directorywizard.com/api
    description: Production server
  - url: https://staging-api.directorywizard.com/api
    description: Staging server
  - url: http://localhost:3001/api
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      summary: Register new user
      description: Create a new user account with validation
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /auth/login:
    post:
      summary: User login
      description: Authenticate user and receive JWT token
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /auth/profile:
    get:
      summary: Get user profile
      description: Retrieve current user profile information
      tags:
        - Authentication
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
        - agreeToTerms
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 8
          example: SecurePassword123!
        firstName:
          type: string
          minLength: 1
          maxLength: 50
          example: John
        lastName:
          type: string
          minLength: 1
          maxLength: 50
          example: Smith
        businessName:
          type: string
          maxLength: 100
          example: ABC Company
        businessRole:
          type: string
          maxLength: 50
          example: Owner
        agreeToTerms:
          type: boolean
          example: true
        subscribeToNewsletter:
          type: boolean
          example: false
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          example: SecurePassword123!
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expiresIn:
          type: string
          example: 24h
    User:
      type: object
      properties:
        id:
          type: string
          example: usr_123456789
        email:
          type: string
          format: email
          example: user@example.com
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Smith
        businessName:
          type: string
          example: ABC Company
        businessRole:
          type: string
          example: Owner
        emailVerified:
          type: boolean
          example: true
        subscribeToNewsletter:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserProfile:
      allOf:
        - $ref: '#/components/schemas/User'
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        message:
          type: string
          example: Detailed error description
        field:
          type: string
          example: email
        code:
          type: string
          example: VALIDATION_ERROR
        timestamp:
          type: string
          format: date-time
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ConflictError:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  retry_after:
                    type: integer
                    description: Seconds to wait before retrying
                    example: 900
...